Refactor Navigation component to use NavLink for improved link handling and consistency
